fix: simplify hosted snapshot restore streaming
